refactor(Crews): 제적 위험자 정렬을 위한 Comparator 구현을 Crews에서 하도록 변경
